Initial try for a vc4 performance monitor
If the hardware is powered down, the counter read 0xdeadbeef, so you
have to cope with shutdowns. Currently this detection is very crude,
omitting a restart of the counters, leading to an endless Nan and Inf
as output.

Maybe we implement a "do not power off" mutex in the kernel. But if
we do this, then we should export the performance counters from the
kernel, too.

Currently runs only as root.

Signed-off-by: Jonas Pfeil <pfeiljonas@gmx.de>